kernel_module_loader: add flag to skip GKI folders
On some devices, parsing all of the GKI folders for modules
to load causes unnecessary delay to the boot process, which
prevents some firmware from loading. The result is that
the battery sysfs entries never get loaded.

Since the modules needed for the device and not typically located
in the GKI folder for the standard kernel, allow skipping these
folders to eliminate the overhead from parsing all of the modules
in these folders.

Usage:
TW_LOAD_VENDOR_MODULES_EXCLUDE_GKI := true

Change-Id: Ia619a5fb0f28ba716bde3972a969c78bfdac1a13